Update install instructions
We now recommend that folks install to `/usr/local/pulumi` on *nix and
we've updated our `install.sh` script so it is the one that is
responsible for creating this folder. If someone has installed to the
old location in `/opt/pulumi` earlier, all they need to do is delete
`/opt/pulumi` and then remove the line from their ~/.profile (assuming
they added added one) that added `/opt/pulumi` to the $PATH

The Windows instructions remain close to their previous behavior with
the modification that the ZIP itself now contains the Pulumi folder,
so it should be extracted to just C:\Program Files.