Overhaul styling
This change overhauls the styling of the docs website fairly
thoroughly as part of pulumi/docs#88.  We have eliminated the
prior Tachyons styling in favor of Material Lite.  I will be
the first to admit that there's a lot of work to be done, but
I think this is good enough to get in now as people as starting
to use the website during the Private Beta.