Fix broken left-nav links
I vaguely recall that Jekyll permits `.html`-less links, while our
custom Go web server does not.  I believe that's the reason the broken
link detector happily succeeds even though we have broken links.

In this case, the new left-nav emitted broken links for the section
links.  The fix is simple: manually append `.html`.

We will want to be on the lookout for more instances of this ...