Move @media CSS styles into a dedicated sheet
This change moves the @media CSS styles associated with the welcome
page into a dedicated stylesheet.  These pertain to the entire document,
and so cannot be used inline, which is one reason the welcome page was
essentially unusable on mobile devices.  It's much better now and looks
almost identical to our product website (with a few minor deltas).

This is a small part of pulumi/docs#148.