Add dashboard.json
To allow updating the Pulumi Console's Dashboard cards faster than going through the normal testing->staging->production deployment of the service, and since these particular cards are docs-focused, we'll host this JSON file on the docs site and have the Console request it from the browser.

For now, News will be a list of recent releases with links to the changelog, but eventually we'll want to change it so it's sourced from the Pulumi blog (via JSON Feed or RSS).