Fix relative CLI links
This fixes the relative CLI links caused by the relative_import of
reference/cli/pulumi.md from reference/commands.md.  The simplest fix
would have obviously been to just edit the reference/commands.md file,
but then it wouldn't be pure generated code.  Instead, I've done
something I'm not proud of, but it seems to be the quickest path to
fixing this given our current situation with Jekyll/Liquid: use a custom
regex to rewrite the specific links that we're targeting on import.