Tighten up docs styles
Our docs feel a bit too "airy" to me.  This is largely due to using
16px font for the text (which is quite large), plus the excessive
line heights.  I also have grown to dislike the -webkit-font-smoothing
we've applied to all of the text (why does this always happen).

This change fixes all of those problems, along with a few other edits:

1) Use Menlo for the code font (similar to GoDocs).

2) Add a slight border, with minimal radius, to code blocks.

3) Lessen the margins for headers, also tightening up things.