Merge Concepts section underneath Reference
Now that we have the new Tutorial/Quickstart => Tour => Reference flow,
having a distinct Concepts in between doesn't make much sense, and just
makes the content feel more crowded than it needs to be.

This new layout feels like a good place to land: the Tour is an
interactive walkthrough that skimps on the details, while Reference is
where you go to get complete details about topics.

I'll admit that the Reference content feels a bit too "toury" to me in
its current form -- and the landing page feels especially "salesy" for
reference content, but let's land the right structure and now we can
focus on evolving the sections as appropriate in the days to come.