Add an API Doc Generation tool
This change includes a simple tool that takes Typedoc files -- generated
from our JSDoc comments in our TypeScript packages -- and creates API
Documentation in Markdown suitable for our docs website.

The bulk of this tool is the interpretation of Typedoc's JSON
representation.  From there, we simply raise the ASTs into a slightly
higher level module representation.  In this, we "cheat" a bit, so that
the documentation is tailored to the idiomatic Pulumi module structure.
This is largely why we can generate better docs than Typedoc does.
(yay, ES6 modules...)

It is, as a result, not infinitely generalizable.

We use Mustache templates so that editing the Markdown results is
reasonable straightforward, although, I must warn, the whitespace
sensitivity is a little cumbersome.  That said, way better than
maintaining YAHWCG (yet-another-hand-written-code-generator).

There are plenty of improvements on their way, but this is a significant
improvement to what we've got -- including deep linking to GitHub!!! --
so I'm going to land now and incrementally refine with feedback in hand.