Make progress on the landing flow
This change introduces a bit of the new landing flow (pulumi/docs#312),
in addition to starting to incorporate some of the new brand colors.

This is a work in progress, but inches us a little closer to what we
think we want, and feels strictly better for active private beta users.