Overhaul the tour
This change totally overhauls the tour's flow.

The new approach is to cover a very simple basic tutorial approach for
the first lesson, focused on code and only the most essential commands,
and then to come back in a second lesson to focus on "teaching" some
of the essential concepts and showing some advanced features.

The flow is complete, although there are 2-3 pages still to be finished.