Add anchor tags to our headers
This uses anchorjs to include anchor tags on all of our header
elements.  You get the nice hover behavior where if you hover over a
header, a little anchor link pops up and you can right click it and
copy the link and then share it.

There were a handfull of places where we explicity used header
elements and we didn't actually want links. In these places, I've
applied the `no-anchor` class and instructed anchorjs not to decorate
headers with this class.

It's possible that there are few other places we'll want to do this,
but from my grepping it looked like this would be sufficent. If we
find more places, we can just assign this class by hand.

Fixes #454