Fix broken "Getting Started" link
This item doesn't have a `section` so it's link was rendering as `href="{{ path }}"` instead of href="{{ path }}.html", causing a broken link because `index` doesn't exist but `index.html` does.

We workaround this by adding a `toplevel` property to indicate it should be rendered as a top-level item that gets a `.html` suffix in its link.