Fix TOC listing for reference docs (#548)
No longer shows the incorrect AWS TOC for References.

This change simplifies how the current TOC is determined based on the
path of the page.

Instead of listing paths to the .md files, list out the path and title
for the pages explicitly. This avoids having to lookup the page in
`site.pages` to get the title and path for the page, which takes forever
for the reference docs, due to how many there are.

The build now runs in 40 seconds, instead of over 5 minutes it would
have taken with the fixed Reference TOC.

Also fixed the FAQ and Known Issues under References to actually show
the selected style when selected.