Fix broken links, run in PR
This change fixes a handful of broken links. I wasn't sure why a
few slipped through recently, when the `make test` target was actually
running clean, and it turns out we had been aggressively filtering out
many links we shouldn't be. This change fixes that, and pulumi/docs#568
tracks fixing the last two remaining categories of known issues.

This also half-fixes pulumi/docs#298, by running the broken link
checker in PR. It isn't a complete fix, because to do that, we'd need
to run the checker against the actual pulumi.io websites after doing
a deployment. In principle, not hard, but will require a bit of plumbing
of stack outputs for the URL into the Makefile `test` target. In
practice, this PR checking should catch most errors, and finds them
sooner during PR when a developer still has an opportunity to fix them.