Add quickstart notes on how to install AWS plugin
Customers sometimes run into issues with the AWS quickstart when their
AWS plugin did not automatically install as part of their package
installation. In the event that NPM did not correctly install their
plugin, this commit adds notes to the relevant areas that instruct users
on how to install the plugin manually.