Elide local paths from type/member links
In some cases, the TypeDoc file will contain absolute paths to
files on the local machine, such as when a type or member resolves
to a standard library name (e.g., the `name` property). In those
cases, we'd previously emit hyperlinks to local filesystem paths.
Instead of doing that, we'll try to fall back to the nearest member
we can resolve or, worst case, just link to the package itself.

Part of pulumi/docs#492.