HTML escape default values
I hadn't realized that the DefaultValue fields for let/const
initializers would contain complex expressions, in the form of
strings. By emitting them as-is, we can end up emitting text like

    let foo = __config.getObject<boolean>("foo");

in the middle of a chunk of HTML. The <boolean> part is then
interpreted as an HTML tag, and it leads to awful <div> mismatches
that garble the layout of the remainder of the page.

The fix here is to simply HTML encode those values in the result.

Fixes pulumi/docs#738.