Add a sign in button to the top nav
It's unfortunate that we don't currently have a sign in button
prominently displayed on our docs website. This is a nice "teaser"
to get you to sign up. This change rectifies this, adding it to the
right of the search box. We don't have the necessary horizontal space
on mobile, however, so it's hidden there.