Merge nodes and modules upon conflicts
This changes our TypeScript doc generation to merge modules and their
constituent nodes upon a name collision. This happens with the way we
do monkey patching of our serverless events, but is a general problem
that can come up with various legal TypeScript patterns.

For now, this simply entails:

    - Treat `declare module { ... }` as being exported; prior to
      this change, we assumed it was internal, and skipped over them.

    - Normalize names like "./x" to being a child of the parent module.

    - Merge conflicting modules.

    - Merge conflicting type nodes. The only legal combinations we
      support right now are interface+interface, class+interface, and
      interface+class. We prefer to choose the class as the "primary"
      for purposes of display (so it shows up as `class Foo` with
      its constructor function, and so on), and generally merge the
      child nodes, interfaces implemented, and source locations.

This fixes pulumi/docs#769. I will re-generate the docs with this
fix in a separate commit after various other changes have landed.