Sort all nodes and children by labels
We sorted the modules and members, but not the children, by their
labels, alphabetically. This means that we already have unsorted
children (like properties) that are out of order! It appears tsdoc
produces its node arrays in a deterministic order, which is why it
was at least stable; but obviously, alphabetically sorted members
is kind of useful. This became even more apparent as we started
merging modules and members, because the result was seemingly random.

The fix is easy: deeply sort all nodes by their labels.