Add Pulumi Python SDK Docs
This PR augments the documentation generator to generate documentation
for the Pulumi package in addition to providers. Instead of
autogenerating all documentation as we do for providers, this commit
opts to autogenerate only the API documentation, while interspersing it
with prose.

In addition, the addition of the Pulumi Python SDK to Sphinx's input
allows Sphinx to intelligently generate cross-reference links for types
such as pulumi.ResourceOptions. There's not a whole lot of cross-linking
happening, but the fact that we're now doing it a little bit
demonstrates that it is possible.