Generate Python Kubernetes documentation (#846)
* Toughen up the Python documentation code generator

The Python documentation code generator works well for packages that
don't have more than one layer of nested submodules. Unfortunately, the
Kubernetes package has many nested submodules and the documentation tool
did not handle them well at all.

This commit is a partial rewrite of the core module rendering logic that
generates Sphinx inputs from provider packages. The logic is basically
the same as before except that it is correct in the face of deeply
nested submodules.

Unfortunately, this revealed a few bugs in tfgen, which will be
addressed in follow-up PRs to pulumi-terraform and this repo. The
documentation generator skips packages that tfgen messed up.

Despite the rewrite, the only documentation diffs are due to the
documentation generator skipping modules that exhibit bugs in tfgen.

* Regenerate documentation

* Enable generation of Kubernetes docs

* Re-generate docs after merge