Always show code examples for Node packages
Code examples in doc comments for our Node packages are written in TypeScript
and use `ts` as the language type, which means they end up being rendered as
`<div class="language-ts">...</div>`.

This means that when JavaScript is selected, these code examples are hidden.

But since they are denoted as `ts` and not `typescript`, even when TypeScript is
selected, the examples are still hidden.

This change ensures these code snippets are always visible, when either
JavaScript or TypeScript are selected.