Ensure markdown is rendered for API docs code examples  (#1113)
Hugo's markdown renderer will not render markdown between HTML tags, and does not support the `markdown="1"` attribute like Jekyll does to enable this. Instead, we have a `md` shortcode that be used around content that should be rendered as markdown. This change fixes our template for Node.js API docs to wrap some markdown content, so that it is rendered correctly, and regenerates all the Node.js docs.