Add proper 301 redirects for our Node.js API docs
With the move to Hugo, the Node.js API docs have moved from `reference/pkg/nodejs/@pulumi/` to `reference/pkg/nodejs/pulumi/` as Hugo strips the `@` when generating the site.

We have Hugo aliases in place that redirect from the old locations to the new locations. However, these are based on [meta refresh](https://www.w3.org/TR/WCAG20-TECHS/H76.html), which does not pass along any anchors that may be specified in the source URL. This is problematic as many of our API docs rely on anchors when linking to specific identifiers.

The fix is to use real 301 redirects instead of these client-side redirects, as browsers will pass along the anchor when 301 redirects are encountered.