Fix strange markdown rendering in API docs
Hugo's markdown renderer does not support rendering markdown inside HTML tags and does not support the `markdown="1"` attribute on HTML tags that Jekyll supports to enable markdown rendering inside the tags.

To work around this limitation, we have a custom `md` shortcode that can be used inside HTML tags to wrap markdown that should be rendered. However, there was a bug in our `md` shortcode. We use the `{{% md %}}` syntax instead of `{{< md >}}` because we intend to have markdown inside the shortcode, and inside our `md` layout, we use `{{ markdownify .Inner }}` to render the `.Inner` content as markdown. However, in some cases, when not actually inside HTML tags, `.Inner` will end up being already rendered markdown. So we end up passing the already rendered markdown through the `markdownify` function, resulting in some wacky rendering.

The fix is to specify the behavior of `{{% %}}` per https://gohugo.io/content-management/shortcodes/#shortcodes-with-markdown, which consistently renders `.Inner` as markdown and emit `.Inner` directly without passing it through `markdownify`.