Use a slightly different flow
Previously, we had the TOC right at the top, and then both the
generic blurb related to the AWS service in question, followed by the
Pulumi Crosswalk for AWS specific bits.

This new change places the generic blurb at the top, right below
the title, followed by the TOC and then the Pulumi Crosswalk for AWS
specific bits in the Overview section. I find that this flows better:

* The generic bit is at the top, and limited to a single paragraph,
  to set context.

* Then you see the TOC which helps glance and go if you're on a mission
  to get something done!

* Next you'll see what Pulumi Crosswalk for AWS specifically adds
  beyond the service itself, and flows into the jobs to be done.