Fix issue preventing use of fixed width font in code snippets (#1163)
We have some typography scss copied over from the www site that defines styles for both the `p` tag as well as a `p` class. This isn't ideal because portions of code snippets are now annotated with a `p` class by Chroma for code highlighting, resulting in the wrong font being used.

The fix is to simply remove the `p` class as it's not needed.