Add fallback og:image
We support setting an og:image when a page has a `meta_image` set in its front matter, but we don't have a fallback, so links to most pulumi.io pages currently don't have an image on Twitter. This fixes that by adding a fallback like the www site (and soon-to-be merged site).