Add a "copy to clipboard" icon to the langchooser
This change adds a copy icon to the upper right of the langchooser,
and copies the selected code body to the clipboard when clicked.

This is part of pulumi/docs#1023. I do think we should explore whether
we can use this in other places -- e.g., none of the API docs samples
will benefit from this because they don't use the langchooser -- but
doing so is a little more complex. We could just look for <div
class="highlight"> elements, but it's not clear, visually, where we'd
put the icon. Perhaps floating with a negative offset in the upper
right would work, but there are also plenty of <div class="highlight">
elements we probably wouldn't want this for.