Update _pygments.scss to select .highlight instead of .chroma
The syntax highlighted code generated by Hugo looks like:

```
<div class="highlight">
    <pre class="chroma">
        <code class="language-foo">
        ...
        </code>
    </pre>
</div>
```

And our styles looks like:

```
.chroma .k { ... }
.chroma .kc { ... }
...
```

This works good, but our generated Python docs, which uses Pygments to do the syntax highlighting, doesn't use `.chroma`, so it doesn't get any syntax highlighting currently, even though all the Pygments classes are there.

This is true for our Node.js docs as well, although, based on how we generate the code for that currently, we could relatively easily update the templates to include a `.chroma` class.

In all cases, the code that we want highlighted will have a `.highlight` class surrounding it, so just update our styles to select `.highlight` instead of `.chroma`.

```
.highlight .k { ... }
.highlight .kc { ... }
...
```