Fix bug in lang chooser
If I've previously viewed a page in our docs with the language tabs,
such as the Get Started guide, and I click the Python tab, that language
choice will be persisted to a cookie.

If I then visit our Node.js API docs, the lang chooser will see the
saved value of "Python" and try to select that tab, hiding all other
language snippets. But our Node.js API docs only contain JavaScript and
TypeScript snippets, so the end result is that all snippets are hidden
and no tab is selected.

This change fixes that. If the value in the cookie does not exist in any
of the tabs on the page, fallback to the default (JavaScript), otherwise
fallback to the first tab that is available on the page.