API doc styling improvements (#1247)
* Update API doc styling for Python

Bring back the styles used for our Python API docs, convert to use Tailwind, and make further tweaks to make it look even better. For example, all definitions now look like code blocks and indentation and spacing is used to improve the hierarchy of the content.

* Only emit the <h1>.Title</h1> if there is a .Title

The generated Python API docs don't currently include front matter and therefore do not have a title. We should fix the Python docs gen to do that, but in the meantime, only emit `<h1>.Title</h1>` if the page has a `.Title`, to avoid an unnecessary gap at the top of the page.

* Update _pygments.scss to select .highlight instead of .chroma

The syntax highlighted code generated by Hugo looks like:

```
<div class="highlight">
    <pre class="chroma">
        <code class="language-foo">
        ...
        </code>
    </pre>
</div>
```

And our styles looks like:

```
.chroma .k { ... }
.chroma .kc { ... }
...
```

This works good, but our generated Python docs, which uses Pygments to do the syntax highlighting, doesn't use `.chroma`, so it doesn't get any syntax highlighting currently, even though all the Pygments classes are there.

This is true for our Node.js docs as well, although, based on how we generate the code for that currently, we could relatively easily update the templates to include a `.chroma` class.

In all cases, the code that we want highlighted will have a `.highlight` class surrounding it, so just update our styles to select `.highlight` instead of `.chroma`.

```
.highlight .k { ... }
.highlight .kc { ... }
...
```

* Update API doc styling for Node.js