More small fixes (#1254)
* Make the header more responsive

Adds padding to medium and above, and adjusts font size and spacing at medium width in order to accommodate the number of items we’re displaying currently. Also uses the full Pulumi logo, since we have room for that now, and adds a little color to the hamburger.

Signed-off-by: Christian Nunciato <c@nunciato.org>

* Bottom-align the image on the whitepaper layout

Also removes the `header` selector from `_hero.scss` in order to reduce the specificity of the rule, allowing it to be overidden more easily with other Tailwind classnames.

Signed-off-by: Christian Nunciato <c@nunciato.org>

* Add team@pulumi link to the footer

Signed-off-by: Christian Nunciato <c@nunciato.org>

* Restore the newsroom and investors sections

Somehow these got dropped! :scream:

Signed-off-by: Christian Nunciato <c@nunciato.org>