Do not use named profiles for deployments
There's an issue today where named profiles which use
`credential_source=Environment` don't actually behave correctly with
the assume-role tool. Instead of actually assuming the role, it just
uses the credentials in the environment, which is not exactly what you
want.

For now, explicitly set the ARN we want to assume, instead of
indirecting across a named profile.