Just use FontAwesome for the icon
Instead of taking a dependency on the Material Icons, which doesn't have all the icons we need elsewhere, just use an icon from FontAwesome.

Also, add a solid background behind the icon, so when the button is over code (when the browser width is very small), it isn't lost above the text.