Friendly error when npm packages aren't installed
Our use of PostCSS in Hugo requires the `postcss-cli` npm package to be installed. If it isn't, Hugo will print an error:

> PostCSS not found; install with "npm install postcss-cli". See https://gohugo.io/hugo-pipes/postcss/

However, this isn't ideal for us, as we use `yarn` in this repo, and don't use it directly. Instead, `make ensure` should be used.

My recent change to add a copy button to code snippets also depends on an npm package (`clipboard-polyfill`) to be installed. If it isn't, Hugo fails with several error messages about not being able to read the file, and it isn't immediately obvious how to address the issue.

This change makes it more clear how to address the problem. If the relevant directories do not exist in `node_modules`, we fail the Hugo build and print an error that says "Run `make ensure`".