Fix broken images on certain blog posts
The problem with these images is that they are above the `<!--more-->` and are therefore included in summary pages such as:

/blog/tag/aws/page/2/easy-eks.png
/blog/tag/kubernetes/page/2/easy-eks.png
/blog/tag/kubernetes/page/2/status-rich.gif
/blog/author/luke-hoban/easy-eks.png
/blog/page/4/easy-eks.png
/blog/page/7/status-rich.gif
/blog/author/alex-clemmer/status-rich.gif

Since the images are relative to the content page, they end up being broken links in these summary pages. If we want to keep these images as page resources (vs. including them in /static/images), we'll need to use a shortcode so we can get a `RelPermalink` via `$.Page.Resources.GetMatch`.