Fix Python docs regression
I'd made some improvements to the styling of the Python API docs before the website merge, but it regressed last week when the structure of the docs changed in a way that the CSS selectors no longer applied.

This fixes the selectors for the new structure, and makes some minor other tweaks while making changes here.