Make broken link checker force follow robot exclusions (#1386)
I noticed our broken link checker was checking the home page and then would complete without checking the rest of the site. It turns out our recent addition of adding a `robots.txt` with `Disallow: *` for non-production builds was preventing the checker from checking links throughout the rest of the site. This change passes the `--follow` flag to the checker, to force it to follow robot exclusions.

It also includes some additional links to exclude, and fixes a couple blog posts to avoid broken images on summary pages.