Fix broken links in @pulumi/postgresql docs
We weren't including `@pulumi/postgresql` in the `gitHubBaseURLs` map, so links to types in that package in the API docs were broken.

Modified the code to return an error if a package is missing from the map, so we don't silently introduce these broken links again in the future with new packages.