Validate the content of the /static/latest-version file
While making related changes here, I've added a test that will ensure no trailing newlines (or other whitespace) is added to the `latest-version` file, which may break dependencies that aren't resilient to it. We recently saw this in a PR and luckily the broken link checker caught it, but this makes the test more explicit.