Cleanup the Changelog (#1458)
* Cleanup the Changelog

- Removed old entries that are being moved to the CHANGELOG.md in `pulumi/pulumi` repo.
- Remove "Changelog" from TOC. The "Download and Install" page links to this page at the bottom.
- Make the links in the table actually look like links.
- Added a `changelog-table-row` shortcode (replacing the `sdk-links` shortcode), which will make it easier to add new rows to the table without introducing subtle broken anchor links.

* Validate the content of the /static/latest-version file

While making related changes here, I've added a test that will ensure no trailing newlines (or other whitespace) is added to the `latest-version` file, which may break dependencies that aren't resilient to it. We recently saw this in a PR and luckily the broken link checker caught it, but this makes the test more explicit.