Add some spacing between sections of the install page
Heading tags that aren't the first child in a container get `mt-8` applied, but we have a few heading tags on the install page as the first child inside divs, so they don't get this margin. This change adds some margin to the divs on the page to add some breathing room between sections.