Use 301 responses for redirects instead of meta refresh redirects (#1606)
* Modernize the Pulumi program

- Upgrade packages (to take subsequent advantage of stronger typing for `routingRules` in more recent versions of `@pulumi/aws`)
- Use Pulumi's built-in TypeScript support -- no need for the manual build step
- Get tslint clean

* Remove some code that is no longer needed

- The existing routing rules applied to the previous pulumi.io site and are no longer needed here.
- Skipping certain dirs/files is also no longer necessary with Hugo's publish directory.

* Use 301 responses for redirects instead of meta refresh tags

When crawling the directory, if the file is an .html file that contains a meta refresh tag, configure the bucket object to be a website redirect that returns a 301 response when requested. We do this to allow #anchors to flow across to the redirect page, which doesn't work with the meta refresh tag -based redirects.