Add a mktutorial tool
This change adds a new mktutorial tool. It takes our examples repo,
gathers up the examples as a series of tutorials, and emits three things:

1) A global index of tutorials, available as the shortcode
   {{< tutorials-index >}}.

2) A cloud index of tutorials, one per cloud, available as the
   shortcode {{< tutorials-index-[cloud] >}}.

3) A page per tutorial.

We currently filter out anything that isn't using the AWS, Azure,
Google Cloud, or Kubernetes cloud providers, and/or anything that
isn't using the TypeScript, JavaScript, or Python language.

This isn't actually wired up and used yet -- that will come shortly.