Prevent smart quotes for string literals in API docs
By default, Hugo's markdown renderer does smart punctuation substitutions, such as turning normal double- and single-quote marks into curly quotes, single hyphens into an ndash and a double hyphens into an mdash, etc.

This is great for most of our content, but there are cases where we don't want this to happen, such as in custom HTML we emit in our Node.js API docs. In particular, we don't want quotes surrounding string literals to be converted to curly quotes.

Aside: In general, any markdown within HTML tags will not render as markdown with Hugo's default markdown processor. However, it appears there are cases where it can happen -- such as with the content inside these custom `<pre>` tags.

To prevent this from happening in our API docs, we wrap a custom `md-disable` shortcode around the custom HTML we don't want to be rendered by the markdown processor, being sure to use the `<` character when referencing the shortcode, which indicates the inner content of the shortcode is already HTML and should not be processed any further.