Shrink the margin on blockquotes slightly (#1617)
* Shrink the margin on blockquotes slightly

The margin on blockquotes, with the orange bar on the left,
looked a bit large to my eyes. This has the effect of making
the blockquote, which is often used as an aside, much more
disruptive to the flow -- particularly for smaller blockquotes
(like single liners). Especially with the smaller font we use
for blockquotes, this "feels" more appropriate.

* Use tailwind style margin instead