Avoid nil pointer dereference in doc generation (#1632)
The doc generation for beta releases of `@pulumi/pulumi` was failing with a nil pointer dereference because some new files were added to the `tsconfig.json`, which included a case we haven't seen before.

This is the type that's causing the issue: https://github.com/pulumi/pulumi/blob/8da252270f7734a8c16d3bd4ce8e208509865d47/sdk/nodejs/runtime/closure/v8_v11andHigher.ts#L99-L101

When generating the type label for type literals, and looping through the node's children, it's possible the `child.Type` node to be nil with this case. Instead, it is a function and needs to be processed as such.

Note: This particular type doesn't actually end up in the docs, but we still process the TypeDoc node for it. This avoids the nil pointer dereference.