Update package constraints for Python docs generation
The docs generation logic for python creates a virtualenv, installs
all our packages into it and uses this as the set of packages to
generate documentation from.

Since a normal `pip install` will not pick up pre-release
packages (like our beta packages), update the constraints for packages
we have released pre-release versions of to pick up the latest versions.