Emit `types` module consistently (#1712)
Sometimes when generating the docs, we'd leave out the `types` module.
Other times, it would be included.

This nondeterministic behavior was due to us modifying a map while
iterating over it. According to the [Go "for statements"
spec](https://golang.org/ref/spec#For_statements):

>  If a map entry is created during iteration, that entry may be
>  produced during the iteration or may be skipped. The choice may vary
>  for each entry created and from one iteration to the next.

This started becoming an issue when we recently started including
`types` module. Previously, all other modules were parented under the
root "index" module, which is always present. But now we have a `types`
module which gets added during processing, and sometimes that entry
would be included and other times it would be skipped during the
iteration.

To address this, we now copy the entries to an array, which will be
processed until it is empty, allowing for items to be added during the
iteration.