Avoid some more 302 redirects
Removes cloud-aws and cloud-azure API docs. These are implementation details that we stopped generating docs for a while ago. But since we don't delete docs we no longer generate, these were sticking around in an old format, with links that do not have a trailing slash. The fix is to "delete" the content and simple redirect to the public `@pulumi/cloud` docs.

Also uses `relref` in one more place so that we have a trailing slash.